autotriage_agent/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
requests
psycopg2-binary
torch
numpy
